Changed the signature of the method to get the time frame (returns int instead of long)
git-svn-id: https://alma-svn.aoc.nrao.edu/p2/trunk/ACS@60829 523d945c-050c-4681-91ec-863ad3bb968a

+++ LGPL/CommonSoftware/jlogEngine/src/com/cosylab/logging/settings/ExpertPrefsDlg.java
@@ -164,7 +164,7 @@ public class ExpertPrefsDlg extends JDialog implements ActionListener {
 	 * 
 	 * @return The time frame selcted in the CB
 	 */
-	public long getTimeFrame() {
+	public int getTimeFrame() {
 		return ((UserPreferences.TimeOption)timeCB.getSelectedItem()).getTimeFrame();
 	}
